move files into buffers folder
