+++ lua/vuffers/buffers/init.lua
@@ -1,0 +1,4 @@
+local bufs = require("vuffers.buffers.buffers")
+local M = vim.tbl_deep_extend("force", {}, bufs)
+
+return M
