+++ lua/vuffers/subscriptions.lua
@@ -2,7 +2,7 @@ local eb = require("vuffers.event-bus")
 local ui = require("vuffers.ui")
 local keymaps = require("vuffers.key-bindings")
 local buffers = require("vuffers.buffers")
-local buffer_actions = require("vuffers.buffer-actions")
+local buffer_actions = require("vuffers.buffers.buffer-actions")
 local config = require("vuffers.config")
 local list = require("utils.list")
 local logger = require("utils.logger")
